I think I want these functions in these places instead
